Added Helper class for controllers
- Added a helper class to avoid having duplicate code for controllers.
- For now, this helper class provides a GetDomainUrl() function that returns the domain name that the api uses.
